Fixed some compiler warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@5999 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/IndigoXMLDoc.h
@@ -45,7 +45,7 @@ public:
 
 	void saveDoc(const std::string& path);
 
-	static const std::string getErrorPositionDesc(const std::string& buffer_str, int offset);
+	static const std::string getErrorPositionDesc(const std::string& buffer_str, ptrdiff_t offset);
 	
 	static void test();
